requirements.txt
fastapi==0.111.1
pandas==2.2.2
sqlalchemy==2.0.31
requests==2.32.3
pyjwt==2.8.0
azure-storage-blob==12.20.0
azure-communication-email==1.0.0
langchain==0.2.8
langchain_community==0.2.7
langchain_openai==0.1.16
loguru==0.7.2
mysql-connector==2.2.9
azure-keyvault==4.2.0
azure-identity==1.19.0